chore: latest generation of Go SDKs

+++ SDKs/amazonaws.com/machinelearning/2014-12-12/go/pkg/models/shared/prediction.go
@@ -5,6 +5,6 @@ package shared
 type Prediction struct {
 	Details         map[string]string  `json:"details,omitempty"`
 	PredictedLabel  *string            `json:"predictedLabel,omitempty"`
-	PredictedScores map[string]float64 `json:"predictedScores,omitempty"`
+	PredictedScores map[string]float32 `json:"predictedScores,omitempty"`
 	PredictedValue  *float32           `json:"predictedValue,omitempty"`
 }
