+++ SDKs/motaword.com/1.0/go/pkg/models/shared/projectinsearch.go
@@ -17,7 +17,7 @@ type ProjectInSearchPriceWithoutDiscount struct {
 }
 
 type ProjectInSearch struct {
-	AverageScores          map[string]float64                   `json:"average_scores,omitempty"`
+	AverageScores          map[string]float32                   `json:"average_scores,omitempty"`
 	BudgetCode             *string                              `json:"budget_code,omitempty"`
 	CallbackURL            *string                              `json:"callback_url,omitempty"`
 	Client                 *User                                `json:"client,omitempty"`
